Move PromptTemplate class to root
